fixes regarding comments
