+++ app/app/components/Root/Root.tsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux'
 import { ConnectedRouter } from 'connected-react-router'
 import { History } from 'history'
 import { MuiThemeProvider } from '@material-ui/core/styles'
-import LP from '../../../provider/language'
-import { translationMessages } from '../../../provider/language/i18n'
+import LP from '../../../language'
+import { translationMessages } from '../../../language/i18n'
 import App from './App'
 import muiTheme from '../../../mui-theme'
