+++ app/language/i18n.ts
@@ -7,9 +7,9 @@
 import { addLocaleData } from 'react-intl'
 import enLocaleData from 'react-intl/locale-data/en'
 
-import { DEFAULT_LOCALE } from './constants'
+import { DEFAULT_LOCALE } from './reducer'
 
-import * as enJson from '../../translations/en.json'
+import * as enJson from '../translations/en.json'
 
 const enTranslationMessages = (<any>enJson).default
