--- app/provider/language/constants.ts
@@ -1,8 +1,0 @@
-/*
- *
- * LanguageProvider constants
- *
- */
-
-export const CHANGE_LOCALE = 'app/LanguageToggle/CHANGE_LOCALE'
-export const DEFAULT_LOCALE = 'en'
